+++ etna/models/base.py
@@ -106,17 +106,17 @@ class Model(ABC, BaseMixin):
         return segment_predict
 
 
-class FitAbstractModel(ABC):
+class AbstractModel(ABC, BaseMixin):
     """Interface for model with fit method."""
 
     @property
     @abstractmethod
     def context_size(self) -> int:
-        """Upper bound to context size of the model."""
+        """Context size of the model. Determines how many history points do we ask to pass to the model."""
         pass
 
     @abstractmethod
-    def fit(self, ts: TSDataset) -> "FitAbstractModel":
+    def fit(self, ts: TSDataset) -> "AbstractModel":
         """Fit model.
 
         Parameters
@@ -151,8 +151,16 @@ class FitAbstractModel(ABC):
         pass
 
 
-class ForecastAbstractModel(ABC):
-    """Interface for model with forecast method."""
+class NonPredictionIntervalContextIgnorantAbstractModel(AbstractModel):
+    """Interface for models that don't support prediction intervals and don't need context for prediction."""
+
+    @property
+    def context_size(self) -> int:
+        """Context size of the model. Determines how many history points do we ask to pass to the model.
+
+        Zero for this model.
+        """
+        return 0
 
     @abstractmethod
     def forecast(self, ts: TSDataset) -> TSDataset:
@@ -171,8 +179,39 @@ class ForecastAbstractModel(ABC):
         pass
 
 
-class PredictIntervalAbstractModel(ABC):
-    """Interface for model with forecast method that creates prediction interval."""
+class NonPredictionIntervalContextRequiredAbstractModel(AbstractModel):
+    """Interface for models that don't support prediction intervals and need context for prediction."""
+
+    @abstractmethod
+    def forecast(self, ts: TSDataset, prediction_size: int) -> TSDataset:
+        """Make predictions.
+
+        Parameters
+        ----------
+        ts:
+            Dataset with features
+        prediction_size:
+            Number of last timestamps to leave after making prediction.
+            Previous timestamps will be used as a context for models that require it.
+
+        Returns
+        -------
+        :
+            Dataset with predictions
+        """
+        pass
+
+
+class PredictionIntervalContextIgnorantAbstractModel(AbstractModel):
+    """Interface for models that support prediction intervals and don't need context for prediction."""
+
+    @property
+    def context_size(self) -> int:
+        """Context size of the model. Determines how many history points do we ask to pass to the model.
+
+        Zero for this model.
+        """
+        return 0
 
     @abstractmethod
     def forecast(
@@ -197,12 +236,153 @@ class PredictIntervalAbstractModel(ABC):
         pass
 
 
-class PerSegmentBaseModel(FitAbstractModel, BaseMixin):
-    """Base class for holding specific models for per-segment prediction."""
+class PredictionIntervalContextRequiredAbstractModel(AbstractModel):
+    """Interface for models that support prediction intervals and need context for prediction."""
+
+    @abstractmethod
+    def forecast(
+        self,
+        ts: TSDataset,
+        prediction_size: int,
+        prediction_interval: bool = False,
+        quantiles: Sequence[float] = (0.025, 0.975),
+    ) -> TSDataset:
+        """Make predictions.
+
+        Parameters
+        ----------
+        ts:
+            Dataset with features
+        prediction_size:
+            Number of last timestamps to leave after making prediction.
+            Previous timestamps will be used as a context for models that require it.
+        prediction_interval:
+            If True returns prediction interval for forecast
+        quantiles:
+            Levels of prediction distribution. By default 2.5% and 97.5% are taken to form a 95% prediction interval
+
+        Returns
+        -------
+        :
+            Dataset with predictions
+        """
+        pass
+
+
+class ModelForecastMixin(ABC):
+    """Base class for model mixins."""
+
+    @abstractmethod
+    def _forecast(self, **kwargs) -> TSDataset:
+        pass
+
+
+class NonPredictionIntervalContextIgnorantModelMixin(ModelForecastMixin):
+    """Mixin for models that don't support prediction intervals and don't need context for prediction."""
+
+    def forecast(self, ts: TSDataset) -> TSDataset:
+        """Make predictions.
+
+        Parameters
+        ----------
+        ts:
+            Dataset with features
+
+        Returns
+        -------
+        :
+            Dataset with predictions
+        """
+        return self._forecast(ts=ts)
+
+
+class NonPredictionIntervalContextRequiredModelMixin(ModelForecastMixin):
+    """Mixin for models that don't support prediction intervals and need context for prediction."""
+
+    def forecast(self, ts: TSDataset, prediction_size: int) -> TSDataset:
+        """Make predictions.
+
+        Parameters
+        ----------
+        ts:
+            Dataset with features
+        prediction_size:
+            Number of last timestamps to leave after making prediction.
+            Previous timestamps will be used as a context for models that require it.
+
+        Returns
+        -------
+        :
+            Dataset with predictions
+        """
+        return self._forecast(ts=ts, prediction_size=prediction_size)
+
+
+class PredictionIntervalContextIgnorantModelMixin(ModelForecastMixin):
+    """Mixin for models that support prediction intervals and don't need context for prediction."""
+
+    def forecast(
+        self, ts: TSDataset, prediction_interval: bool = False, quantiles: Sequence[float] = (0.025, 0.975)
+    ) -> TSDataset:
+        """Make predictions.
+
+        Parameters
+        ----------
+        ts:
+            Dataset with features
+        prediction_interval:
+            If True returns prediction interval for forecast
+        quantiles:
+            Levels of prediction distribution. By default 2.5% and 97.5% are taken to form a 95% prediction interval
+
+        Returns
+        -------
+        :
+            Dataset with predictions
+        """
+        return self._forecast(ts=ts, prediction_interval=prediction_interval, quantiles=quantiles)
+
+
+class PredictionIntervalContextRequiredModelMixin(ModelForecastMixin):
+    """Mixin for models that support prediction intervals and need context for prediction."""
+
+    def forecast(
+        self,
+        ts: TSDataset,
+        prediction_size: int,
+        prediction_interval: bool = False,
+        quantiles: Sequence[float] = (0.025, 0.975),
+    ) -> TSDataset:
+        """Make predictions.
+
+        Parameters
+        ----------
+        ts:
+            Dataset with features
+        prediction_size:
+            Number of last timestamps to leave after making prediction.
+            Previous timestamps will be used as a context for models that require it.
+        prediction_interval:
+            If True returns prediction interval for forecast
+        quantiles:
+            Levels of prediction distribution. By default 2.5% and 97.5% are taken to form a 95% prediction interval
+
+        Returns
+        -------
+        :
+            Dataset with predictions
+        """
+        return self._forecast(
+            ts=ts, prediction_size=prediction_size, prediction_interval=prediction_interval, quantiles=quantiles
+        )
+
+
+class PerSegmentModelMixin(ModelForecastMixin):
+    """Mixin for holding methods for per-segment prediction."""
 
     def __init__(self, base_model: Any):
         """
-        Init PerSegmentBaseModel.
+        Init PerSegmentModelMixin.
 
         Parameters
         ----------
@@ -213,7 +393,7 @@ class PerSegmentBaseModel(FitAbstractModel, BaseMixin):
         self._models: Optional[Dict[str, Any]] = None
 
     @log_decorator
-    def fit(self, ts: TSDataset) -> "PerSegmentBaseModel":
+    def fit(self, ts: TSDataset) -> "PerSegmentModelMixin":
         """Fit model.
 
         Parameters
@@ -285,29 +465,15 @@ class PerSegmentBaseModel(FitAbstractModel, BaseMixin):
         segment_predict["timestamp"] = dates
         return segment_predict
 
-
-class PerSegmentModel(PerSegmentBaseModel, ForecastAbstractModel):
-    """Class for holding specific models for per-segment prediction."""
-
-    def __init__(self, base_model: Any):
-        """
-        Init PerSegmentBaseModel.
-
-        Parameters
-        ----------
-        base_model:
-            Internal model which will be used to forecast segments, expected to have fit/predict interface
-        """
-        super().__init__(base_model=base_model)
-
     @log_decorator
-    def forecast(self, ts: TSDataset) -> TSDataset:
+    def _forecast(self, ts: TSDataset, **kwargs) -> TSDataset:
         """Make predictions.
 
         Parameters
         ----------
         ts:
             Dataframe with features
+
         Returns
         -------
         :
@@ -315,7 +481,8 @@ class PerSegmentModel(PerSegmentBaseModel, ForecastAbstractModel):
         """
         result_list = list()
         for segment, model in self._get_model().items():
-            segment_predict = self._forecast_segment(model=model, segment=segment, ts=ts)
+            segment_predict = self._forecast_segment(model=model, segment=segment, ts=ts, **kwargs)
+
             result_list.append(segment_predict)
 
         result_df = pd.concat(result_list, ignore_index=True)
@@ -330,61 +497,8 @@ class PerSegmentModel(PerSegmentBaseModel, ForecastAbstractModel):
         return ts
 
 
-class PerSegmentPredictionIntervalModel(PerSegmentBaseModel, PredictIntervalAbstractModel):
-    """Class for holding specific models for per-segment prediction which are able to build prediction intervals."""
-
-    def __init__(self, base_model: Any):
-        """
-        Init PerSegmentPredictionIntervalModel.
-
-        Parameters
-        ----------
-        base_model:
-            Internal model which will be used to forecast segments, expected to have fit/predict interface
-        """
-        super().__init__(base_model=base_model)
-
-    @log_decorator
-    def forecast(
-        self, ts: TSDataset, prediction_interval: bool = False, quantiles: Sequence[float] = (0.025, 0.975)
-    ) -> TSDataset:
-        """Make predictions.
-
-        Parameters
-        ----------
-        ts:
-            Dataset with features
-        prediction_interval:
-            If True returns prediction interval for forecast
-        quantiles:
-            Levels of prediction distribution. By default 2.5% and 97.5% are taken to form a 95% prediction interval
-
-        Returns
-        -------
-        :
-            Dataset with predictions
-        """
-        result_list = list()
-        for segment, model in self._get_model().items():
-            segment_predict = self._forecast_segment(
-                model=model, segment=segment, ts=ts, prediction_interval=prediction_interval, quantiles=quantiles
-            )
-            result_list.append(segment_predict)
-
-        result_df = pd.concat(result_list, ignore_index=True)
-        result_df = result_df.set_index(["timestamp", "segment"])
-        df = ts.to_pandas(flatten=True)
-        df = df.set_index(["timestamp", "segment"])
-        df = df.combine_first(result_df).reset_index()
-
-        df = TSDataset.to_dataset(df)
-        ts.df = df
-        ts.inverse_transform()
-        return ts
-
-
-class MultiSegmentModel(FitAbstractModel, ForecastAbstractModel, BaseMixin):
-    """Class for holding specific models for per-segment prediction."""
+class MultiSegmentModelMixin(ModelForecastMixin):
+    """Mixin for holding methods for multi-segment prediction."""
 
     def __init__(self, base_model: Any):
         """
@@ -398,7 +512,7 @@ class MultiSegmentModel(FitAbstractModel, ForecastAbstractModel, BaseMixin):
         self._base_model = base_model
 
     @log_decorator
-    def fit(self, ts: TSDataset) -> "MultiSegmentModel":
+    def fit(self, ts: TSDataset) -> "MultiSegmentModelMixin":
         """Fit model.
 
         Parameters
@@ -418,7 +532,7 @@ class MultiSegmentModel(FitAbstractModel, ForecastAbstractModel, BaseMixin):
         return self
 
     @log_decorator
-    def forecast(self, ts: TSDataset) -> TSDataset:
+    def _forecast(self, ts: TSDataset, **kwargs) -> TSDataset:
         """Make predictions.
 
         Parameters
@@ -433,7 +547,7 @@ class MultiSegmentModel(FitAbstractModel, ForecastAbstractModel, BaseMixin):
         """
         horizon = len(ts.df)
         x = ts.to_pandas(flatten=True).drop(["segment"], axis=1)
-        y = self._base_model.predict(x).reshape(-1, horizon).T
+        y = self._base_model.predict(x, **kwargs).reshape(-1, horizon).T
         ts.loc[:, pd.IndexSlice[:, "target"]] = y
         ts.inverse_transform()
         return ts
@@ -514,24 +628,6 @@ class DeepAbstractNet(ABC):
 
 class DeepBaseAbstractModel(ABC):
     """Interface for holding class of etna native deep models."""
-
-    @abstractmethod
-    def forecast(self, ts: TSDataset, horizon: int) -> TSDataset:
-        """Make predictions.
-
-        Parameters
-        ----------
-        ts:
-            Dataset with features and expected decoder length for context
-        horizon:
-            Horizon to predict for
-
-        Returns
-        -------
-        :
-            Dataset with predictions
-        """
-        pass
 
     @abstractmethod
     def raw_fit(self, torch_dataset: "Dataset") -> "DeepBaseAbstractModel":
@@ -619,7 +715,7 @@ class DeepBaseNet(DeepAbstractNet, LightningModule):
         return loss
 
 
-class DeepBaseModel(FitAbstractModel, DeepBaseAbstractModel, BaseMixin):
+class DeepBaseModel(DeepBaseAbstractModel, NonPredictionIntervalContextRequiredAbstractModel):
     """Class for partially implemented interfaces for holding deep models."""
 
     def __init__(
@@ -790,15 +886,16 @@ class DeepBaseModel(FitAbstractModel, DeepBaseAbstractModel, BaseMixin):
         return predictions_dict
 
     @log_decorator
-    def forecast(self, ts: "TSDataset", horizon: int) -> "TSDataset":
+    def forecast(self, ts: "TSDataset", prediction_size: int) -> "TSDataset":
         """Make predictions.
 
         Parameters
         ----------
         ts:
             Dataset with features and expected decoder length for context
-        horizon:
-            Horizon to predict for
+        prediction_size:
+            Number of last timestamps to leave after making prediction.
+            Previous timestamps will be used as a context.
 
         Returns
         -------
@@ -812,9 +909,9 @@ class DeepBaseModel(FitAbstractModel, DeepBaseAbstractModel, BaseMixin):
             dropna=False,
         )
         predictions = self.raw_predict(test_dataset)
-        future_ts = ts.tsdataset_idx_slice(start_idx=self.encoder_length, end_idx=self.encoder_length + horizon)
+        future_ts = ts.tsdataset_idx_slice(start_idx=self.encoder_length, end_idx=self.encoder_length + prediction_size)
         for (segment, feature_nm), value in predictions.items():
-            future_ts.df.loc[:, pd.IndexSlice[segment, feature_nm]] = value[:horizon, :]
+            future_ts.df.loc[:, pd.IndexSlice[segment, feature_nm]] = value[:prediction_size, :]
 
         future_ts.inverse_transform()
 
@@ -831,31 +928,19 @@ class DeepBaseModel(FitAbstractModel, DeepBaseAbstractModel, BaseMixin):
         return self.net
 
 
-class MultiSegmentPredictionIntervalModel(FitAbstractModel, PredictIntervalAbstractModel, BaseMixin):
-    """Class for holding specific models for multi-segment prediction which are able to build prediction intervals."""
+ModelType = Union[
+    NonPredictionIntervalContextIgnorantAbstractModel,
+    NonPredictionIntervalContextRequiredAbstractModel,
+    PredictionIntervalContextIgnorantAbstractModel,
+    PredictionIntervalContextRequiredAbstractModel,
+]
 
-    def __init__(self):
-        """Init MultiSegmentPredictionIntervalModel."""
-        self.model = None
+ContextRequiredModelType = Union[
+    NonPredictionIntervalContextRequiredAbstractModel,
+    PredictionIntervalContextRequiredAbstractModel,
+]
 
-    def get_model(self) -> Any:
-        """Get internal model that is used inside etna class.
-
-        Internal model is a model that is used inside etna to forecast segments,
-        e.g. :py:class:`catboost.CatBoostRegressor` or :py:class:`sklearn.linear_model.Ridge`.
-
-        Returns
-        -------
-        :
-           Internal model
-        """
-        return self.model
-
-
-BaseModel = Union[
-    PerSegmentModel,
-    PerSegmentPredictionIntervalModel,
-    MultiSegmentModel,
-    DeepBaseModel,
-    MultiSegmentPredictionIntervalModel,
+ContextIgnorantModelType = Union[
+    NonPredictionIntervalContextIgnorantAbstractModel,
+    PredictionIntervalContextIgnorantAbstractModel,
 ]