+++ etna/models/deadline_ma.py
@@ -6,7 +6,9 @@ from typing import List
 import numpy as np
 import pandas as pd
 
-from etna.models.base import PerSegmentModel
+from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
+from etna.models.base import NonPredictionIntervalContextIgnorantModelMixin
+from etna.models.base import PerSegmentModelMixin
 
 
 class SeasonalityMode(Enum):
@@ -155,7 +157,11 @@ class _DeadlineMovingAverageModel:
         return cur_value
 
 
-class DeadlineMovingAverageModel(PerSegmentModel):
+class DeadlineMovingAverageModel(
+    PerSegmentModelMixin,
+    NonPredictionIntervalContextIgnorantModelMixin,
+    NonPredictionIntervalContextIgnorantAbstractModel,
+):
     """Moving average model that uses exact previous dates to predict."""
 
     def __init__(self, window: int = 3, seasonality: str = "month"):