+++ etna/models/prophet.py
@@ -10,7 +10,9 @@ import pandas as pd
 
 from etna import SETTINGS
 from etna.models.base import BaseAdapter
-from etna.models.base import PerSegmentPredictionIntervalModel
+from etna.models.base import PerSegmentModelMixin
+from etna.models.base import PredictionIntervalContextIgnorantAbstractModel
+from etna.models.base import PredictionIntervalContextIgnorantModelMixin
 
 if SETTINGS.prophet_required:
     from prophet import Prophet
@@ -153,7 +155,9 @@ class _ProphetAdapter(BaseAdapter):
         return self.model
 
 
-class ProphetModel(PerSegmentPredictionIntervalModel):
+class ProphetModel(
+    PerSegmentModelMixin, PredictionIntervalContextIgnorantModelMixin, PredictionIntervalContextIgnorantAbstractModel
+):
     """Class for holding Prophet model.
 
     Notes
@@ -197,8 +201,6 @@ class ProphetModel(PerSegmentPredictionIntervalModel):
     2020-04-15      5.00      7.00      4.00      7.00
     2020-04-16      8.00      6.00      2.00      0.00
     """
-
-    context_size = 0
 
     def __init__(
         self,