+++ etna/models/sarimax.py
@@ -13,7 +13,9 @@ from statsmodels.tsa.statespace.sarimax import SARIMAXResultsWrapper
 
 from etna.libs.pmdarima_utils import seasonal_prediction_with_confidence
 from etna.models.base import BaseAdapter
-from etna.models.base import PerSegmentPredictionIntervalModel
+from etna.models.base import PerSegmentModelMixin
+from etna.models.base import PredictionIntervalContextIgnorantAbstractModel
+from etna.models.base import PredictionIntervalContextIgnorantModelMixin
 from etna.models.utils import determine_num_steps
 
 warnings.filterwarnings(
@@ -352,7 +354,9 @@ class _SARIMAXAdapter(_SARIMAXBaseAdapter):
         return result
 
 
-class SARIMAXModel(PerSegmentPredictionIntervalModel):
+class SARIMAXModel(
+    PerSegmentModelMixin, PredictionIntervalContextIgnorantModelMixin, PredictionIntervalContextIgnorantAbstractModel
+):
     """
     Class for holding Sarimax model.
 
@@ -363,8 +367,6 @@ class SARIMAXModel(PerSegmentPredictionIntervalModel):
     additional features what is not known in future, and regressors for features we do know in
     future.
     """
-
-    context_size = 0
 
     def __init__(
         self,