+++ etna/models/seasonal_ma.py
@@ -5,7 +5,9 @@ from typing import List
 import numpy as np
 import pandas as pd
 
-from etna.models.base import PerSegmentModel
+from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
+from etna.models.base import NonPredictionIntervalContextIgnorantModelMixin
+from etna.models.base import PerSegmentModelMixin
 
 
 class _SeasonalMovingAverageModel:
@@ -92,7 +94,11 @@ class _SeasonalMovingAverageModel:
         return y_pred
 
 
-class SeasonalMovingAverageModel(PerSegmentModel):
+class SeasonalMovingAverageModel(
+    PerSegmentModelMixin,
+    NonPredictionIntervalContextIgnorantModelMixin,
+    NonPredictionIntervalContextIgnorantAbstractModel,
+):
     """
     Seasonal moving average.