+++ etna/models/sklearn.py
@@ -6,8 +6,10 @@ import pandas as pd
 from sklearn.base import RegressorMixin
 
 from etna.models.base import BaseAdapter
-from etna.models.base import MultiSegmentModel
-from etna.models.base import PerSegmentModel
+from etna.models.base import MultiSegmentModelMixin
+from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
+from etna.models.base import NonPredictionIntervalContextIgnorantModelMixin
+from etna.models.base import PerSegmentModelMixin
 
 
 class _SklearnAdapter(BaseAdapter):
@@ -72,10 +74,12 @@ class _SklearnAdapter(BaseAdapter):
         return self.model
 
 
-class SklearnPerSegmentModel(PerSegmentModel):
+class SklearnPerSegmentModel(
+    PerSegmentModelMixin,
+    NonPredictionIntervalContextIgnorantModelMixin,
+    NonPredictionIntervalContextIgnorantAbstractModel,
+):
     """Class for holding per segment Sklearn model."""
-
-    context_size = 0
 
     def __init__(self, regressor: RegressorMixin):
         """
@@ -89,10 +93,12 @@ class SklearnPerSegmentModel(PerSegmentModel):
         super().__init__(base_model=_SklearnAdapter(regressor=regressor))
 
 
-class SklearnMultiSegmentModel(MultiSegmentModel):
+class SklearnMultiSegmentModel(
+    MultiSegmentModelMixin,
+    NonPredictionIntervalContextIgnorantModelMixin,
+    NonPredictionIntervalContextIgnorantAbstractModel,
+):
     """Class for holding Sklearn model for all segments."""
-
-    context_size = 0
 
     def __init__(self, regressor: RegressorMixin):
         """